+++ android/src/main/java/io/rownd/android/models/RowndHubInteropMessage.kt
@@ -34,6 +34,8 @@ enum class MessageType {
     CreatePasskey,
     @SerialName("trigger_sign_in_with_passkey")
     AuthenticateWithPasskey,
+    @SerialName("hub_loaded")
+    HubLoaded,
     @SerialName("hub_resize")
     HubResize,
     @SerialName("can_touch_background_to_dismiss")
@@ -77,6 +79,7 @@ data class TriggerSignInWithGoogleMessage(
 data class TriggerSignInWithGooglePayload(
     @SerialName("intent")
     var intent: RowndSignInIntent? = null,
+    var hint: String? = null,
 )
 
 @Serializable
@@ -105,6 +108,10 @@ data class HubResizePayload(
     @SerialName("height")
     var height: String? = null,
 )
+@Serializable
+data class HubLoaded(
+    override var type: MessageType = MessageType.HubLoaded
+) : RowndHubInteropMessage()
 
 @Serializable
 data class HubResizeMessage(
@@ -113,16 +120,16 @@ data class HubResizeMessage(
 ) : RowndHubInteropMessage()
 
 @Serializable
-data class CanTouchBackgroundToDismissPayload(
-    @SerialName("enable")
-    var enable: String? = null,
-)
-
-@Serializable
 data class CanTouchBackgroundToDismissMessage(
     override var type: MessageType = MessageType.CanTouchBackgroundToDismiss,
     var payload: CanTouchBackgroundToDismissPayload
 ) : RowndHubInteropMessage()
+
+@Serializable
+data class CanTouchBackgroundToDismissPayload(
+    @SerialName("enable")
+    var enable: String
+)
 
 object RowndHubInteropMessageSerializer : JsonContentPolymorphicSerializer<RowndHubInteropMessage>(RowndHubInteropMessage::class) {
     override fun selectDeserializer(element: JsonElement): DeserializationStrategy<out RowndHubInteropMessage> {
@@ -135,6 +142,7 @@ object RowndHubInteropMessageSerializer : JsonContentPolymorphicSerializer<Rownd
             "close_hub_view_controller" -> CloseHubViewMessage.serializer()
             "trigger_sign_up_with_passkey" -> SignUpWithPasskeyMessage.serializer()
             "trigger_sign_in_with_passkey" -> SignInWithPasskeyMessage.serializer()
+            "hub_loaded" -> HubLoaded.serializer()
             "hub_resize" -> HubResizeMessage.serializer()
             "can_touch_background_to_dismiss" -> CanTouchBackgroundToDismissMessage.serializer()
             else -> throw Error("Key '$messageType' did not match a known serializer.")