+++ android/src/main/java/io/rownd/android/views/ComposableBottomSheet.kt
@@ -29,6 +29,7 @@ import com.airbnb.lottie.compose.LottieCompositionSpec
 import com.airbnb.lottie.compose.LottieConstants
 import com.airbnb.lottie.compose.rememberLottieComposition
 import io.rownd.android.Rownd
+import io.rownd.android.util.convertStringToColor
 import kotlinx.coroutines.launch
 
 import androidx.compose.runtime.saveable.rememberSaveable
@@ -102,6 +103,8 @@ abstract class ComposableBottomSheetFragment : DialogFragment() {
             //  when certain operations are in progress.
             // confirmStateChange = { false }
         )
+
+
 
         SideEffect {
             sheetState = bottomSheetState