+++ android/src/main/java/io/rownd/android/views/key_transfer/CameraView.kt
@@ -1,37 +1,26 @@
 package io.rownd.android.views.key_transfer
 
 import android.content.Context
-import android.net.Uri
-import android.util.Log
-import androidx.camera.core.*
+import androidx.camera.core.CameraSelector
+import androidx.camera.core.ExperimentalGetImage
+import androidx.camera.core.ImageCapture
+import androidx.camera.core.Preview
 import androidx.camera.lifecycle.ProcessCameraProvider
 import androidx.camera.view.PreviewView
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.sharp.Add
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.core.content.ContextCompat
-import java.io.File
-import java.util.concurrent.Executor
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 
@@ -44,6 +33,7 @@ private suspend fun Context.getCameraProvider(): ProcessCameraProvider = suspend
 }
 
 @Composable
+@ExperimentalGetImage
 fun CameraView(
 //    outputDirectory: File,
 //    executor: Executor,
@@ -63,7 +53,6 @@ fun CameraView(
         .requireLensFacing(lensFacing)
         .build()
 
-    @ExperimentalGetImage
     val imageAnalysis = setupAnalysisUseCase(
         onSuccess = onValueCaptured,
         onFailure = onError